// src/omnifocus/types.ts
export interface Context {
  "persistentIdentifier"?: string 
  "allowsNextAction": number 
  "altitude": number | null 
  "availableTaskCount": number 
  "childrenCount": number 
  "childrenCountActive": number 
  "childrenState": number 
  "containedTaskCount": number 
  "creationOrdinal": number 
  "dateAdded": Date 
  "dateHidden": Date | null 
  "dateModified": Date 
  "effectiveDateHidden": Date | null 
  "latitude": number | null 
  "localNumberOfDueSoonTasks": number 
  "localNumberOfOverdueTasks": number 
  "locationName": string | null 
  "longitude": number | null 
  "name": string | null 
  "nextTaskCount": number 
  "noteXMLData": any | null 
  "notificationFlags": number | null 
  "parent": string | null 
  "plainTextNote": string | null 
  "radius": number | null 
  "rank": number 
  "remainingTaskCount": number 
  "tasksAreUserOrdered": number 
  "totalNumberOfDueSoonTasks": number 
  "totalNumberOfOverdueTasks": number 
}
// ...
export interface ProjectInfo {
  "pk"?: string 
  "containsSingletonActions": number 
  "effectiveStatus": string 
  "folder": string | null 
  "folderEffectiveDateHidden": Date | null 
  "lastReviewDate": Date | null 
  "minimumDueDate": Date | null 
  "nextReviewDate": Date | null 
  "nextTask": string | null 
  "numberOfAvailableTasks": number 
  "numberOfContainedTasks": number 
  "numberOfDueSoonTasks": number 
  "numberOfOverdueTasks": number 
  "numberOfRemainingTasks": number 
  "reviewRepetitionString": string | null 
  "status": string 
  "task": string | null 
  "taskBlocked": number 
  "taskBlockedByFutureStartDate": number 
  "taskDateToStart": Date | null 
}
export interface Setting {
  "persistentIdentifier"?: string 
  "creationOrdinal": number 
  "dateAdded": Date 
  "dateModified": Date 
  "valueData": any | null 
}
export interface Task {
  "persistentIdentifier"?: string 
  "blocked": number 
  "blockedByFutureStartDate": number 
  "childrenCount": number 
  "childrenCountActive": number 
  "childrenCountAvailable": number 
  "childrenCountCompleted": number 
  "childrenState": number 
  "completeWhenChildrenComplete": number 
  "containingProjectContainsSingletons": number 
  "containingProjectInfo": string | null 
  "containsNextTask": number 
  "creationOrdinal": number 
  "dateAdded": Date 
  "dateCompleted": Date | null 
  "dateDue": Date | null 
  "dateHidden": Date | null 
  "dateModified": Date 
  "dateToStart": Date | null 
  "deferDateAlarmPolicyString": string | null 
  "dueDateAlarmPolicyString": string | null 
  "dueSoon": number 
  "effectiveContainingProjectInfoActive": number 
  "effectiveContainingProjectInfoRemaining": number 
  "effectiveDateCompleted": Date | null 
  "effectiveDateDue": Date | null 
  "effectiveDateHidden": Date | null 
  "effectiveDateToStart": Date | null 
  "effectiveFlagged": number 
  "effectiveInInbox": number 
  "estimatedMinutes": number | null 
  "flagged": boolean 
  "hasCompletedDescendant": number 
  "hasFlaggedTaskInTree": number 
  "hasUnestimatedLeafTaskInTree": number 
  "inInbox": boolean 
  "latestTimeToStartAlarmPolicyString": string | null 
  "maximumEstimateInTree": number | null 
  "minimumEstimateInTree": number | null 
  "name": string | null 
  "nextCloneIdentifier": number 
  "nextTaskOfProjectInfo": string | null 
  "noteXMLData": any | null 
  "overdue": number 
  "parent": string | null 
  "plainTextNote": string | null 
  "projectInfo": string | null 
  "rank": number 
  "repetitionMethodString": string | null 
  "repetitionRuleString": string | null 
  "sequential": number 
  "tagged": number 
}
export interface TaskToTag {
  "persistentIdentifier"?: string 
  "creationOrdinal": number 
  "dateAdded": Date 
  "dateModified": Date 
  "rankInTag": any | null 
  "rankInTask": any 
  "tag": string | null 
  "task": string | null 
}